fixed a collsion bug and added a ton of soap related stuff

+++ src/map/MapTile.java
@@ -3,16 +3,16 @@ package map;
 public class MapTile {
 	//Container class for map tiles
 	public TileData properties;
-	public int x;
-	public int y;
+	public double x;
+	public double y;
 	public TileEntitiy partner;
-	public MapTile (TileData tileId, int x, int y) {
+	public MapTile (TileData tileId, double x, double y) {
 		//tileId is in the format [tile name]
 		this.properties = tileId;
 		this.x = x;
 		this.y = y;
 	}
-	public MapTile (TileData tileId, int x, int y, TileEntitiy partner) {
+	public MapTile (TileData tileId, double x, double y, TileEntitiy partner) {
 		//tileId is in the format [tile name]
 		this.properties = tileId;
 		this.x = x;
